# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Meeting Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface Meeting {
  id: number;
  projectId: number;
  title: string;
  date: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface MeetingInput {
  /** @minLength 1 */
  title: string;
  date: string;
  notes?: string;
}

export interface MeetingUpdateInput {
  /** @minLength 1 */
  title?: string;
  date?: string;
  /** @nullable */
  notes?: string | null;
}

export interface ProjectUpdateInput {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
}

export interface ProjectSummary {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  meetingCount: number;
  /** @nullable */
  latestMeetingDate?: string | null;
  createdAt: string;
}

export interface RecentMeeting {
  id: number;
  projectId: number;
  title: string;
  date: string;
  projectName: string;
}

export interface Stats {
  totalProjects: number;
  totalMeetings: number;
  recentMeetings: RecentMeeting[];
}

export interface SearchProjectResult {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface SearchMeetingResult {
  id: number;
  projectId: number;
  projectName: string;
  title: string;
  date: string;
  /** @nullable */
  snippet?: string | null;
}

export interface SearchResult {
  projects: SearchProjectResult[];
  meetings: SearchMeetingResult[];
}

export interface TimelineEntry {
  meetingId: number;
  title: string;
  date: string;
  highlight: string;
  /** @nullable */
  changeFromPrevious?: string | null;
}

export interface ProjectTimeline {
  entries: TimelineEntry[];
}

export type MaterialType = typeof MaterialType[keyof typeof MaterialType];


export const MaterialType = {
  photo: 'photo',
  image: 'image',
  pdf: 'pdf',
  excel: 'excel',
  pptx: 'pptx',
  text: 'text',
  audio: 'audio',
} as const;

export type MaterialStatus = typeof MaterialStatus[keyof typeof MaterialStatus];


export const MaterialStatus = {
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;

/**
 * @nullable
 */
export type MaterialSpeakerMap = {[key: string]: string} | null;

export interface Material {
  id: number;
  meetingId: number;
  type: MaterialType;
  filename: string;
  originalName: string;
  /** @nullable */
  extractedText?: string | null;
  /** @nullable */
  contextNote?: string | null;
  /** @nullable */
  speakerMap?: MaterialSpeakerMap;
  status: MaterialStatus;
  createdAt: string;
}

export type SpeakerMapInputSpeakerMap = {[key: string]: string};

export interface SpeakerMapInput {
  speakerMap: SpeakerMapInputSpeakerMap;
}

export interface MaterialFileInput {
  type: MaterialType;
  file: Blob;
}

export type MaterialTextInputType = typeof MaterialTextInputType[keyof typeof MaterialTextInputType];


export const MaterialTextInputType = {
  text: 'text',
} as const;

export interface MaterialTextInput {
  type: MaterialTextInputType;
  name?: string;
  content: string;
  contextNote?: string;
}

export type SearchParams = {
/**
 * @minLength 1
 */
q: string;
};