# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meeting Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/meetings/{meetingId}/materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List materials for a meeting
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
        - name: meetingId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
        "404":
          description: Project or meeting not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Upload a file or text material to a meeting
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
        - name: meetingId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MaterialFileInput"
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialTextInput"
      responses:
        "201":
          description: Created material (may still be processing)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project or meeting not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/meetings/{meetingId}/materials/{materialId}:
    patch:
      operationId: updateMaterialSpeakers
      tags: [materials]
      summary: Assign real names to diarized speakers in an audio transcript
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
        - name: meetingId
          in: path
          required: true
          schema:
            type: number
        - name: materialId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakerMapInput"
      responses:
        "200":
          description: Updated material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
        - name: meetingId
          in: path
          required: true
          schema:
            type: number
        - name: materialId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Material deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/meetings/{meetingId}/materials/{materialId}/retry:
    post:
      operationId: retryMaterial
      tags: [materials]
      summary: Retry text extraction for a failed material
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
        - name: meetingId
          in: path
          required: true
          schema:
            type: number
        - name: materialId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Material reset to processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    MaterialType:
      type: string
      enum: [photo, image, pdf, excel, pptx, text, audio]
    MaterialStatus:
      type: string
      enum: [processing, ready, error]
    Material:
      type: object
      required: [id, meetingId, type, filename, originalName, status, createdAt]
      properties:
        id:
          type: number
        meetingId:
          type: number
        type:
          $ref: "#/components/schemas/MaterialType"
        filename:
          type: string
        originalName:
          type: string
        extractedText:
          type: ["string", "null"]
        contextNote:
          type: ["string", "null"]
        speakerMap:
          type: ["object", "null"]
          additionalProperties:
            type: string
        status:
          $ref: "#/components/schemas/MaterialStatus"
        createdAt:
          type: string
          format: date-time
    SpeakerMapInput:
      type: object
      required: [speakerMap]
      properties:
        speakerMap:
          type: object
          additionalProperties:
            type: string
    MaterialFileInput:
      type: object
      required: [type, file]
      properties:
        type:
          $ref: "#/components/schemas/MaterialType"
        file:
          type: string
          format: binary
    MaterialTextInput:
      type: object
      required: [type, content]
      properties:
        type:
          type: string
          enum: [text]
        name:
          type: string
        content:
          type: string
        contextNote:
          type: string